+++ packages/ws-worker/test/api/workloop.test.ts
@@ -6,7 +6,6 @@ import { mockChannel } from '../../src/mock/sockets';
 import startWorkloop from '../../src/api/workloop';
 import { CLAIM } from '../../src/events';
 import { createMockLogger } from '@openfn/logger';
-import { execute } from '../../src/api/execute';
 
 let cancel;
 
@@ -19,15 +18,18 @@ test.afterEach(() => {
 test('workloop can be cancelled', async (t) => {
   let count = 0;
   let cancel;
-  const channel = mockChannel({
-    [CLAIM]: () => {
-      count++;
-      cancel();
-      return { attempts: [] };
-    },
-  });
+  const app = {
+    channel: mockChannel({
+      [CLAIM]: () => {
+        count++;
+        cancel();
+        return { attempts: [] };
+      },
+    }),
+    execute: () => {},
+  };
 
-  cancel = startWorkloop(channel, () => {}, logger, { timeout: 1 });
+  cancel = startWorkloop(app, logger, 1, 1);
 
   await sleep(100);
   // A quirk of how cancel works is that the loop will be called a few times
@@ -37,14 +39,19 @@ test('workloop can be cancelled', async (t) => {
 test('workloop sends the attempts:claim event', (t) => {
   return new Promise((done) => {
     let cancel;
-    const channel = mockChannel({
-      [CLAIM]: () => {
-        t.pass();
-        done();
-        return { attempts: [] };
-      },
-    });
-    cancel = startWorkloop(channel, () => {}, logger);
+
+    const app = {
+      workflows: {},
+      channel: mockChannel({
+        [CLAIM]: () => {
+          t.pass();
+          done();
+          return { attempts: [] };
+        },
+      }),
+      execute: () => {},
+    };
+    cancel = startWorkloop(app, logger, 1, 1);
   });
 });
 
@@ -52,34 +59,42 @@ test('workloop sends the attempts:claim event several times ', (t) => {
   return new Promise((done) => {
     let cancel;
     let count = 0;
-    const channel = mockChannel({
-      [CLAIM]: () => {
-        count++;
-        if (count === 5) {
-          t.pass();
-          done();
-        }
-        return { attempts: [] };
-      },
-    });
-    cancel = startWorkloop(channel, () => {}, logger);
+    const app = {
+      workflows: {},
+      channel: mockChannel({
+        [CLAIM]: () => {
+          count++;
+          if (count === 5) {
+            t.pass();
+            done();
+          }
+          return { attempts: [] };
+        },
+      }),
+      execute: () => {},
+    };
+    cancel = startWorkloop(app, logger, 1, 1);
   });
 });
 
 test('workloop calls execute if attempts:claim returns attempts', (t) => {
   return new Promise((done) => {
     let cancel;
-    const channel = mockChannel({
-      [CLAIM]: () => ({
-        attempts: [{ id: 'a', token: 'x.y.z' }],
-      }),
-    });
 
-    const execute = (attempt) => {
-      t.deepEqual(attempt, { id: 'a', token: 'x.y.z' });
-      t.pass();
-      done();
+    const app = {
+      workflows: {},
+      channel: mockChannel({
+        [CLAIM]: () => ({
+          attempts: [{ id: 'a', token: 'x.y.z' }],
+        }),
+      }),
+      execute: (attempt) => {
+        t.deepEqual(attempt, { id: 'a', token: 'x.y.z' });
+        t.pass();
+        done();
+      },
     };
-    cancel = startWorkloop(channel, execute, logger);
+
+    cancel = startWorkloop(app, logger, 1, 1);
   });
 });