+++ packages/lightning-mock/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/lightning-mock
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies [ac7b0ca]
+  - @openfn/engine-multi@0.1.4
+
 ## 1.0.4
 
 ### Patch Changes